Renamed an API folder
