+++ etc/webapi.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <routes xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../../../../../app/code/Magento/Webapi/etc/webapi.xsd">
-    <route url="/v1/charge" method="POST">
+    <route url="/v1/payment/charge" method="GET">
         <service class="CheckoutCom\Magento2\Api\PaymentInterface" method="charge"/>
         <resources>
             <resource ref="anonymous"/>
